crawler/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
trafilatura==1.9.0
tenacity==8.3.0
psycopg2-binary==2.9.9
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
tqdm==4.66.4
loguru==0.7.2
fake-useragent==1.5.1
urllib3==2.2.2
